stmback/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-guardian==2.4.0
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
protobuf==4.21.12
six==1.16.0
sqlparse==0.4.4
